Add multi-instrument (BTC) trading, live risk safeguards, and Telegram alerts
Major features:
- AI chat overhaul with WhatsApp-style UI, tool-calling direction bias, and
  optimistic sending; today's P&L badge and direction-bias badge in header
- 3-way theme picker (dark/light/warm); settings restructured, chart-controls
  popup removed in favor of the Settings page
- Emotion avatar (reactive robot face on trade win/loss), Shilo strategy
  (EMA50 breakout), backtest script for session-restricted scalping analysis
- Half-target and full-target breakeven stop management on open trades
- EMA-50 trend filter (block counter-trend entries) and post-close cooldown
- Daily profit-target auto-stop, now unified under a persisted auto-stop
  reason so a page reload still explains why the bot is off
- Multi-instrument trading (added BTC_USD) with per-instrument thought/EMA50
  filtering, float-based position sizing for small-contract instruments
- Telegram notifications on trade open/close (opt-in via .env)

Bug fixes found and fixed via live verification against the real MT5 account:
- Strategy engine now retries on startup if MT5 hands back stale seed
  history, and trips the kill switch (rather than trading blind) if it stays
  stale — a real instance of this let two SELL trades through against the
  EMA50 trend filter
- Fixed the broker reconnecting to the .env default symbol instead of the
  persisted active instrument on every restart
- Fixed breakeven-extension sizing using the (possibly already-moved) current
  stop instead of the trade's original risk
- Chart EMA50/entry-price-line axis-label collision; mobile centering bug in
  DraggableWindow; test isolation from the real Telegram bot

+++ backend/app/broker/base.py
@@ -48,7 +48,7 @@ class BrokerTrade:
     broker_trade_id: str
     instrument: str
     side: Side
-    units: int
+    units: float
     entry_price: float
     unrealized_pnl: float
     stop_loss: Optional[float]
@@ -77,7 +77,7 @@ class BrokerAdapter(ABC):
         self,
         instrument: str,
         side: Side,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
