+++ backend/app/broker/simulated.py
@@ -33,7 +33,7 @@ class SimPosition:
     id: str
     instrument: str
     side: Side
-    units: int
+    units: float
     entry_price: float
     stop_loss: Optional[float]
     take_profit: Optional[float]
@@ -98,7 +98,7 @@ class SimulatedBrokerAdapter(BrokerAdapter):
         self,
         instrument: str,
         side: Side,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
