+++ backend/app/broker/twelvedata.py
@@ -154,7 +154,7 @@ class TwelveDataAdapter(BrokerAdapter):
         self,
         instrument: str,
         side: Side,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
