+++ backend/app/broker/yahoo.py
@@ -113,7 +113,7 @@ class YahooFinanceAdapter(BrokerAdapter):
         self,
         instrument: str,
         side: Side,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
